feat(spot-volume): show spot-volume and market cap (#1318)
